final crawl1.php and crawl_function1.php

+++ dependencies/crawl_functions1.php
@@ -2,20 +2,18 @@
 
 //require  '/var/www/d78236gbe27823/vendor/autoload.php';
 //use thiagoalessio\TesseractOCR\TesseractOCR;
-
-// $eol = "\n";
-// if ($_REQUEST['browser'] == "Yes")
-$eol = PHP_EOL;
+if (php_sapi_name() == "cli") $eol = "\n";
+else  $eol = "<br>";
 
 function filenamedate($epapercode, $conn)
 {
-    // $finddateq = "Select * from Crawl_Record1 WHERE Papershortname='" . $epapercode . "' ORDER BY Paperdate DESC LIMIT 1";
-    // $finddaters = mysqli_query($conn, $finddateq);
-    // if (mysqli_num_rows($finddaters)) {
-    //     $finddaterow = mysqli_fetch_array($finddaters);
-    //     $filedate = date('Y-m-d', strtotime($finddaterow['Paperdate']) + (24 * 3600));
-    // } else
-    $filedate = date('Y-m-d', time());
+    $finddateq = "Select * from Crawl_Record1 WHERE Papershortname='" . $epapercode . "' ORDER BY Paperdate DESC LIMIT 1";
+    $finddaters = mysqli_query($conn, $finddateq);
+    if (mysqli_num_rows($finddaters)) {
+        $finddaterow = mysqli_fetch_array($finddaters);
+        $filedate = date('Y-m-d', strtotime($finddaterow['Paperdate']) + (24 * 3600));
+    } else
+        $filedate = date('Y-m-d', time());
 
     return $filedate;
 }
@@ -26,7 +24,6 @@ function dateForLinks($epapercode, $filenamedate)
     else if ($epapercode == "HB") return date('Y/m/d', strtotime($filenamedate));
     else if ($epapercode == "TOI" or $epapercode == "ET" or $epapercode == "MT" or $epapercode == "Mirror") return  date('d/m/Y', strtotime($filenamedate));
     else if ($epapercode == "GSM") return date("d-m-Y", strtotime($filenamedate));
-
     else if ($epapercode == "DJ" or $epapercode == "NB" or $epapercode == "ND" or $epapercode == "NVR" or $epapercode == "PAP") return date('d-M-Y', strtotime($filenamedate));
     else if ($epapercode == "JPS") return date('dmy', strtotime($filenamedate));
     else if ($epapercode == "NYB" or $epapercode == "RS" or $epapercode == "SAM" or $epapercode == "SMJ") return date('dmY', strtotime($filenamedate));
@@ -195,18 +192,15 @@ function cityArray($epapercode)
             return array("Mangalore", "Davangere", "Kalaburgi", "Hubli", "Bangalore");
             break;
 
-        case "GSM":
-            return array("ahmedabad", "baroda", "bhavnagar", "bhuj", "mumbai", "rajkot", "surat");
-            break;
-
         case "VV":
             return array("Bengaluru", "Hubli");
             break;
-
+        case "GSM":
+            return array("ahmedabad", "baroda", "bhavnagar", "bhuj", "mumbai", "rajkot", "surat");
+            break;
         case "PN":
             return  array("Mumbai", "Pune", "Nashik", "Nagpur", "Kolhapur", "Satara", "Solapur", "Jalgaon", "Dhule", "Nanded", "Thane", "Latur", "Ahmednagar");
             break;
-
         case "TOI":
             return  array("Ahmedabad", "Bangalore", "Bhopal", "Chandigarh", "Chennai", "Delhi", "Goa", "Hyderabad", "Jaipur", "Kochi", "Kolkata", "Lucknow", "Mumbai");
             break;
@@ -219,8 +213,6 @@ function cityArray($epapercode)
         case "Mirror":
             return  array("Bangalore", "Mumbai", "Pune");
             break;
-        default:
-            return null;
     }
 }
 function cityCodeArray($epapercode)
@@ -292,12 +284,9 @@ function cityCodeArray($epapercode)
 
 function makefilepath($epapercode, $city, $date, $number, $lang)
 {
-    $filepath = "./nvme/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".jpg";
+    $filepath = "/nvme/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".jpg";
     $temp_txtfile = str_replace(".jpg", "", $filepath);
-    $txtfile = "./ocrtexts/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".txt";
-    // $filepath = "/nvme/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".jpg";
-    // $temp_txtfile = str_replace(".jpg", "", $filepath);
-    // $txtfile = "/var/www/d78236gbe27823/marketing/Whatsapp/images/ocrtexts/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".txt";
+    $txtfile = "/var/www/d78236gbe27823/marketing/Whatsapp/images/ocrtexts/" . $epapercode . "_" . $city . "_" . $date . "_" . $number . "_admin_" . $lang . ".txt";
     $newspaper_name = $epapercode;
     $newspaper_region = $city;
     $newspaper_date = $date;
@@ -334,7 +323,7 @@ function writeImage($url, $path)
     fclose($handle);
 }
 
-function runTesseract($edition, $page, $section, $conn, $patharray, $lang)
+function runTesseract($epapername, $edition, $page, $section, $conn, $patharray, $lang)
 {
     global $eol;
     $filepath = $patharray[0];
@@ -344,11 +333,12 @@ function runTesseract($edition, $page, $section, $conn, $patharray, $lang)
     $newspaper_region = $patharray[4];
     $newspaper_date = $patharray[5];
     $newspaper_lang = $patharray[6];
-    $newspaper_full_name = $patharray[7];
+    $newspaper_full_name = $epapername;
     $newspaper_operator_name = $patharray[8];
     $starttime = date('Y-m-d H:i:s', time());
 
     try {
+
         $command = "tesseract " . $filepath . " " . $temp_txtfile . " -l " . $lang . " > /dev/null 2>&1";
         exec($command);
         $text = file_get_contents($temp_txtfile . ".txt");
@@ -362,7 +352,7 @@ function runTesseract($edition, $page, $section, $conn, $patharray, $lang)
         $n = count($matches);
 
         if ($n == 0) {
-            echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>Tesseract Completed. No new numbers found" .  $eol;
+            echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>Tesseract Completed. No new numbers found" . $eol;
         } else {
 
             echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>Tesseract Completed. " . $n . " new numbers found. File Saved" . $eol;
@@ -373,39 +363,39 @@ function runTesseract($edition, $page, $section, $conn, $patharray, $lang)
 
             rename($temp_txtfile . ".txt", $txtfile);
 
-            //     echo $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Starting to add in the database...";
+            echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Starting to add in the database...";
 
-            //     $values = "";
+            $values = "";
 
-            //     for ($i = 0; $i < $n; $i++) {
+            for ($i = 0; $i < $n; $i++) {
 
-            //         $blockcheck = "select * from Blocked_Numbers where Mobile_No = '" . $matches[$i] . "'";
-            //         $bcrs = mysqli_query($conn, $blockcheck);
-            //         if (!mysqli_num_rows($bcrs)) {
-            //             $values .= "('" . $matches[$i] . "','" . $newspaper_name . "','" . $newspaper_region . "','" . $newspaper_date . "','" . $newspaper_lang . "','" . $newspaper_full_name . "','" . $newspaper_operator_name . "'),";
-            //         } else echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Skipping " . $matches[$i] . " found in blocked numbers";
-            //     }
+                $blockcheck = "select * from Blocked_Numbers where Mobile_No = '" . $matches[$i] . "'";
+                $bcrs = mysqli_query($conn, $blockcheck);
+                if (!mysqli_num_rows($bcrs)) {
+                    $values .= "('" . $matches[$i] . "','" . $newspaper_name . "','" . $newspaper_region . "','" . $newspaper_date . "','" . $newspaper_lang . "','" . $newspaper_full_name . "','" . $newspaper_operator_name . "'),";
+                } else echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Skipping " . $matches[$i] . " found in blocked numbers";
+            }
 
-            //     if (strlen($values) > 0) {
+            if (strlen($values) > 0) {
 
-            //         $values = substr($values, 0, strlen($values) - 1) . " ON DUPLICATE KEY UPDATE Newspaper_Name = concat(Newspaper_Name,' | ',VALUES(Newspaper_Name)), Newspaper_Region = concat(Newspaper_Region,' | ',VALUES(Newspaper_Region)), Newspaper_Date = concat(Newspaper_Date,' | ',VALUES(Newspaper_Date)), Newspaper_Lang = Newspaper_Lang;";
+                $values = substr($values, 0, strlen($values) - 1) . " ON DUPLICATE KEY UPDATE Newspaper_Name = concat(Newspaper_Name,' | ',VALUES(Newspaper_Name)), Newspaper_Region = concat(Newspaper_Region,' | ',VALUES(Newspaper_Region)), Newspaper_Date = concat(Newspaper_Date,' | ',VALUES(Newspaper_Date)), Newspaper_Lang = Newspaper_Lang;";
 
-            //         $q = "insert into Mobile_Lists1 (Mobile_Number,Newspaper_Name,Newspaper_Region,Newspaper_Date,Newspaper_Lang,Image_File_Name,Image_Operator) values " . $values;
+                $q = "insert into Mobile_Lists1 (Mobile_Number,Newspaper_Name,Newspaper_Region,Newspaper_Date,Newspaper_Lang,Image_File_Name,Image_Operator) values " . $values;
 
-            //         if (!mysqli_query($conn, $q)) {
-            //             echo  $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Error in insert query... ABORTING!" . $eol . $q . "" . $eol;
-            //             die();
-            //         }
+                if (!mysqli_query($conn, $q)) {
+                    echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Error in insert query... ABORTING!" . $eol . $q . "" . $eol;
+                    die();
+                }
 
-            //         echo  $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Insert query executed successfully......" . $eol;
-            //     } else echo  $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "No numbers left to insert";
+                echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "Insert query executed successfully......" . $eol;
+            } else echo "" . $eol . date('Y-m-d H:i:s', (time() + (5.5 * 3600))) . "==> " . "No numbers left to insert";
         }
 
-        // $iq = "INSERT INTO Crawled_Pages1 (Papername,Papershortname,Paperdate,Edition,Page,Section,No_Of_Mobiles_Found,Start_Time) VALUES ('" . $newspaper_full_name . "','" . $newspaper_name . "','" . $newspaper_date . "','" . $edition . "','" . $page . "','" . $section . "','" . count($matches) . "','" . $starttime . "')";
+        $iq = "INSERT INTO Crawled_Pages1 (Papername,Papershortname,Paperdate,Edition,Page,Section,No_Of_Mobiles_Found,Start_Time) VALUES ('" . $newspaper_full_name . "','" . $newspaper_name . "','" . $newspaper_date . "','" . $edition . "','" . $page . "','" . $section . "','" . count($matches) . "','" . $starttime . "')";
 
-        // echo $eol . $iq . "" . $eol;
+        echo "" . $eol . $iq . "" . $eol;
 
-        // mysqli_query($conn, $iq);
+        mysqli_query($conn, $iq);
     } catch (Exception $e) {
         echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>Tesseract Falied to run" . $eol;
     }
@@ -423,9 +413,15 @@ function getHBeditionlink($city, $dateforlinks, $citylink, $code)
     return $link;
 }
 
-function crawltoi($cityarray, $dateForLinks, $epapercode, $citycode, $filenamedate, $eol, $conn, $lang)
+function crawltoi($cityarray, $dateForLinks, $epapercode, $citycode, $filenamedate, $eol, $conn, $lang, $cities_of_interest, $epapername)
 {
     for ($edition = 0; $edition < count($cityarray); $edition++) {
+
+        if (!in_array(ucfirst(explode("-", $cityarray[$edition])[0]), $cities_of_interest)) {
+
+            echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>Skipping " . $cityarray[$edition] . " Edition. Doesn't fall in cities of interest" . $eol;
+            continue;
+        }
 
         $failedPageCount = 0;
         $date_formatted = date("Y/d/m", strtotime($dateForLinks));
@@ -456,28 +452,25 @@ function crawltoi($cityarray, $dateForLinks, $epapercode, $citycode, $filenameda
                 $section_size_array = getimagesize($imagelink);
                 $width = $section_size_array[0];
 
+
                 if ($width > 0) {
                     $imageFound = "Yes";
                     $failedPageCount = 0;
 
                     if ($width > 200 and $width < 250) {
                         $getpath = explode("&", makefilepath($epapercode,  $cityarray[$edition], $filenamedate, $page . "00" . $section, $lang));
-
-                        // if (alreadyDone($getpath[0], $conn) == "Yes") continue;
-
+                        if (alreadyDone($getpath[0], $conn) == "Yes") continue;
                         writeImage($imagelink, $getpath[0]);
 
                         echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>File " . $getpath[0] . " Saved" . $eol;
-                        runTesseract($cityarray[$edition], $page, $section, $conn, $getpath, $lang);
+                        runTesseract($epapername, $cityarray[$edition], $page, $section, $conn, $getpath, $lang);
                         echo date('Y-m-d H:i:s', time() + (5.5 * 3600)) . "=>" . $cityarray[$edition] . " Page " . $page . " Section " . $section . " Completed" . $eol;
                         ob_flush();
                         flush();
-                    }
+                    } else continue;
                 } else {
                     continue;
                 }
-                ob_flush();
-                flush();
             }
 
             if ($imageFound == "No") $failedPageCount++;
